refactor(habitat): move verify model into module
