+++ tools/habitat-harness/src/service/modules/verify/contract.ts
@@ -1,11 +1,11 @@
 import { habitatServiceErrorMap } from "@internal/habitat-harness/service/errors";
 import { CheckReportSchema } from "@internal/habitat-harness/service/model/check/policy/structural/schema";
 import { VerifyTargetPlanSchema } from "@internal/habitat-harness/service/model/workspace/index";
-import { VerifyReceiptSchema } from "@internal/habitat-harness/service/model/verify/dto/verify.schema";
 import type { HabitatServiceProcedureContract } from "@internal/habitat-harness/service/procedure-contract";
 import { toStandardSchema } from "@internal/habitat-harness/service/typebox-standard-schema";
 import { eoc } from "effect-orpc";
 import { type Static, Type } from "typebox";
+import { VerifyReceiptSchema } from "./model/dto/verify.schema.js";
 
 const VerifyServiceRunInputSchema = Type.Object(
   {
