+++ tools/habitat-harness/src/service/modules/verify/model/dto/verify.schema.ts
